Code refactor to make code more manageable
